NewStuff NOTHING WORKS YET

+++ netchangenew/netchangenew/Server.cs
@@ -29,7 +29,7 @@ namespace netchangenew
                 clientOut.AutoFlush = true;
 
                 // De server weet niet wat de poort is van de client die verbinding maakt, de client geeft dus als onderdeel van het protocol als eerst een bericht met zijn poort
-                int zijnPoort = int.Parse(clientIn.ReadLine().Split()[1]);
+                ushort zijnPoort = ushort.Parse(clientIn.ReadLine().Split()[1]);
 
                 Console.WriteLine("Client maakt verbinding: " + zijnPoort);
